osc-pccs: prefetch sqlite3 prebuild
sqlite3 npm package use native bindings which are prebuilt
and made available part of the release.

npm install uses the "prebuild" tool to get the artifacts downloaded
but this fails in the hermetic build environment (hermeto prefetch for
npm does not seem to be able to handle this case).

prefetch the prebuilt native bindings using the generic prefetcher
and point "prebuild" the place where they can be found.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>